requirements.txt
asgiref==3.4.1
Django==3.2.16
django-cors-headers==3.10.1
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.1.1
